Feat: fix protected routes and Todo model

+++ App/Models/TodoModel.php
@@ -6,17 +6,26 @@ namespace App\Models;
 
 use App\AppCore\Model\Model;
 use App\AppCore\Routing\Url;
+use App\AppCore\Utils\Persist;
 use App\Routes\Routes;
 use PDO;
 
 class TodoModel extends Model
 {
+    private int $userId;
+
+    public function __construct(PDO $connection)
+    {
+        parent::__construct($connection);
+
+        $this->userId = Persist::get('loggedUser') ? Persist::get('loggedUser')->id : 0;
+    }
+
     public function getAllTodos()
     {
         try {
-            // TODO: SELECT * FROM `todos` WHERE `user_id`=? ORDER BY `id` DESC
-            $statement = $this->connection->prepare('SELECT * FROM `todos` ORDER BY `id` DESC');
-            $statement->execute();
+            $statement = $this->connection->prepare('SELECT * FROM `todos` WHERE `user_id`=? ORDER BY `id` DESC');
+            $statement->execute([$this->userId]);
             return $statement->fetchAll(PDO::FETCH_OBJ);
         } catch(\PDOException $e) {
             Url::redirect(Routes::AppError);
@@ -26,9 +35,8 @@ class TodoModel extends Model
     public function getTodo(int $id): object|false
     {
         try {
-            // TODO: SELECT * FROM `todos` WHERE `id`=? AND `user_id`=?
-            $statement = $this->connection->prepare('SELECT * FROM `todos` WHERE `id`=?');
-            $statement->execute([$id]);
+            $statement = $this->connection->prepare('SELECT * FROM `todos` WHERE `id`=? AND `user_id`=?');
+            $statement->execute([$id, $this->userId]);
             return $statement->fetch(PDO::FETCH_OBJ);
         } catch(\PDOException $e) {
             Url::redirect(Routes::AppError);
@@ -38,9 +46,8 @@ class TodoModel extends Model
     public function delete(int $id): void
     {
         try {
-            // TODO: DELETE FROM `todos` WHERE `id`=? AND `user_id`=?
-            $statement = $this->connection->prepare('DELETE FROM `todos` WHERE `id`=?');
-            $statement->execute([$id]);
+            $statement = $this->connection->prepare('DELETE FROM `todos` WHERE `id`=? AND `user_id`=?');
+            $statement->execute([$id, $this->userId]);
         } catch(\PDOException $e) {
             Url::redirect(Routes::AppError);
         }
@@ -49,9 +56,8 @@ class TodoModel extends Model
     public function setDone(int $id): void
     {
         try {
-            // TODO: UPDATE `todos` SET `done`=1 WHERE `id`=? AND `user_id`=?
-            $statement = $this->connection->prepare('UPDATE `todos` SET `done`=1 WHERE `id`=?');
-            $statement->execute([$id]);
+            $statement = $this->connection->prepare('UPDATE `todos` SET `done`=1 WHERE `id`=? AND `user_id`=?');
+            $statement->execute([$id, $this->userId]);
         } catch(\PDOException $e) {
             Url::redirect(Routes::AppError);
         }
@@ -60,9 +66,8 @@ class TodoModel extends Model
     public function updateTodo(int $id, string $content): void
     {
         try {
-            // TODO: UPDATE `todos` SET `content`=? WHERE `id`=? AND `user_id`=?
-            $statement = $this->connection->prepare('UPDATE `todos` SET `content`=? WHERE `id`=?');
-            $statement->execute([$content, $id]);
+            $statement = $this->connection->prepare('UPDATE `todos` SET `content`=? WHERE `id`=? AND `user_id`=?');
+            $statement->execute([$content, $id, $this->userId]);
         } catch(\PDOException $e) {
             Url::redirect(Routes::AppError);
         }
@@ -72,8 +77,7 @@ class TodoModel extends Model
     {
         try {
             $statement = $this->connection->prepare('INSERT INTO `todos`(`user_id`, `content`) VALUES (?, ?)');
-            // TODO: použít ID přihlášeného uživatele
-            $statement->execute([1, $content]);
+            $statement->execute([$this->userId, $content]);
         } catch(\PDOException $e) {
             Url::redirect(Routes::AppError);
         }
